requirements.txt
attrs==25.3.0
backports.tarfile==1.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==8.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
MarkupSafe==3.0.2
more-itertools==10.7.0
narwhals==1.43.0
numpy==2.2.6
packaging==24.2
pandas==2.3.0
piexif==1.1.3
pillow==11.2.1
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pyperclip==1.9.0
pypiwin32==223
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6-WebEngine==6.9.0
PyQt6-WebEngine-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
tenacity==9.1.2
toml==0.10.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
win10toast==0.9
zipp==3.23.0
